requirements.txt
pdfplumber
pandas
streamlit
google-api-python-client
google-auth
pytest
